refactor(cli-integ): split refactor dry-run tests into individual files
Split the combined `cdk-refactor-dry-run.integtest.ts` file into three
separate test files to enable parallel execution during integration
testing.
